refactor: test constant 분리
