+++ src/constants/test.js
@@ -1,0 +1,13 @@
+export const TEST = {
+  MOCK_DATA: {
+    ID: 'test',
+    NAME: '테스트',
+  },
+  ID: {
+    SIGN_UP_BUTTON: 'signup-button',
+    SIGN_IN_BUTTON: 'signin-button',
+    STATION_ADD_BUTTON: 'station-add-button',
+    LINE_ADD_BUTTON: 'line-add-button',
+    SECTION_ADD_BUTTON: 'section-add-button',
+  },
+};
