+++ src/pages/StationManager/index.js
@@ -1,7 +1,8 @@
 import { Formik } from 'formik';
 import React from 'react';
 import { PageTemplate, Input, Button, ManagementList } from '../../components';
-import { ROUTE, COLOR, INPUT_TEXT, TEST } from '../../constants';
+import { ROUTE, COLOR, INPUT_TEXT } from '../../constants';
+import { TEST } from '../../constants/test';
 import { useStationManager } from '../../hooks';
 import { validateStationName } from '../../utils';
 import { Form, InputWrapper, ButtonWrapper, Validator } from './style';
